Learned this morning that NNU can't use INCLUDEALL at all, so 1. NNU and INCLUDEALL lines no longer count towards seat count 2. If INCLUDEALL is used in an options file with an NNU license, you'll see a warning stating you can't use the 2. This is a warning since the license could be using another LO and I don't check for the presense of other LOs.
I've removed the logic that makes INCLUDEALL skip subtraction for the same product on the same license. The reason for this is because I'm using the same philosophy for subtracting CN seats between INCLUDE and INCLUDEALL: if the USER or GROUP has access to it, then we can reasonably predict how many seats may be used. When we use INCLUDEALL, that means everything is given access, so everything should be subtracted.